FPGA V13
'- RFNotReady signal reconverted to RFOnState. Inversion removed from
gpio_interface.vhd and logic corrected in mainprogram
- Timing constraints set back to original
- 3HzTrigDiag debugging added
- PhaseRampUp and PhaseRampDw modified to make sure the phase ramping
always follows the shorter path between phase at beginning of the ramp
and phase at the end of the ramp
- Tuning reset: number of pulses are equalt to '0' when the movement
enable of a plunger is not active. in this way when you re-enable the
movement of a plunger, it will not move until you press TuningReset
button